Moved file specific operations to new class
